Organize reducers by dashboard.
