+++ client/src/reducer/dos/contestFetchOk.ts
