+++ client/src/reducer/dos/fetchDosAsmStateOk.ts
