+++ client/src/reducer/dos/uploadRandomSeedOk.ts
