+++ client/src/reducer/root.ts
@@ -1,25 +1,27 @@
 import * as _ from 'lodash';
 
-import countyDashboardRefreshOk from './countyDashboardRefreshOk';
-import countyFetchAllCvrsOk from './countyFetchAllCvrsOk';
-import countyFetchContestsOk from './countyFetchContestsOk';
-import countyFetchCvrOk from './countyFetchCvrOk';
-import countyLoginOk from './countyLoginOk';
-import dosContestFetchOk from './dosContestFetchOk';
-import dosDashboardRefreshOk from './dosDashboardRefreshOk';
-import dosLoginOk from './dosLoginOk';
-import fetchAuditBoardAsmStateOk from './fetchAuditBoardAsmStateOk';
-import fetchCountyAsmStateOk from './fetchCountyAsmStateOk';
-import fetchCvrsToAuditOk from './fetchCvrsToAuditOk';
-import fetchDosAsmStateOk from './fetchDosAsmStateOk';
+import countyDashboardRefreshOk from './county/dashboardRefreshOk';
+import countyFetchAllCvrsOk from './county/fetchAllCvrsOk';
+import fetchAuditBoardAsmStateOk from './county/fetchAuditBoardAsmStateOk';
+import countyFetchContestsOk from './county/fetchContestsOk';
+import fetchCountyAsmStateOk from './county/fetchCountyAsmStateOk';
+import countyFetchCvrOk from './county/fetchCvrOk';
+import fetchCvrsToAuditOk from './county/fetchCvrsToAuditOk';
+import countyLoginOk from './county/loginOk';
+import updateAcvrForm from './county/updateAcvrForm';
+import uploadAcvrOk from './county/uploadAcvrOk';
+import uploadBallotManifestOk from './county/uploadBallotManifestOk';
+import uploadCvrExportOk from './county/uploadCvrExportOk';
+
+import dosContestFetchOk from './dos/contestFetchOk';
+import dosDashboardRefreshOk from './dos/dashboardRefreshOk';
+import fetchDosAsmStateOk from './dos/fetchDosAsmStateOk';
+import dosLoginOk from './dos/loginOk';
+import selectContestsForAuditOk from './dos/selectContestsForAuditOk';
+import setRiskLimitOk from './dos/setRiskLimitOk';
+import uploadRandomSeedOk from './dos/uploadRandomSeedOk';
+
 import login1FOk from './login1FOk';
-import selectContestsForAuditOk from './selectContestsForAuditOk';
-import setRiskLimitOk from './setRiskLimitOk';
-import updateAcvrForm from './updateAcvrForm';
-import uploadAcvrOk from './uploadAcvrOk';
-import uploadBallotManifestOk from './uploadBallotManifestOk';
-import uploadCvrExportOk from './uploadCvrExportOk';
-import uploadRandomSeedOk from './uploadRandomSeedOk';
 
 
 interface AppState {
